Separated trust editor from pairing wizard into separate widget

+++ integration/v2/src/libs/libid/name/GenderGenerator.cpp
@@ -22,8 +22,8 @@ QString GenderGenerator::generate()
 	OC_METHODGATE();
 	QString gender="Genderless";
 	if(nullptr!=rng) {
-		const int sz=alternatives.size();
-		const int index=(static_cast<int>(rng->generateInt32())) % sz;
+		const quint32 sz=static_cast<quint32>(alternatives.size());
+		const int index=(static_cast<int>(rng->generateInt32()% sz));
 		gender=alternatives[index];
 	}
 	return gender;
